Project1-parent/Project1-biz/src/main/java/org/dao/UserDao.java
/*
 * Copyright (C) 2018 Fastjrun, Inc. All Rights Reserved.
 */
package org.dao;

import org.apache.ibatis.annotations.Insert;
import org.apache.ibatis.annotations.Options;
import org.apache.ibatis.annotations.Update;

import org.entity.User;

public interface UserDao {

    @Insert("insert into t_user(`loginPwd`,`nickName`,`sex`,`mobileNo`,`loginErrCount`,`lastLoginTime`,`createTime`,"
            + "`loginName`,`lastModifyTime`,`email`) values(#{loginPwd},#{nickName},#{sex},#{mobileNo},"
            + "#{loginErrCount},#{lastLoginTime},#{createTime},#{loginName},#{lastModifyTime},#{email})")
    @Options(useGeneratedKeys = true, keyProperty = "id")
    public int insertNoStatusAndLastRecordLoginErrTime(User user);

    @Update("update t_user set `loginPwd` = #{loginPwd},`nickName` = #{nickName},`sex` = #{sex},`mobileNo` = "
            + "#{mobileNo},`loginName` = #{loginName},`email` = #{email} where id = #{id} ")
    public int updateUserInfoById(User user);

}
